refactor(hocuspocus): separate library exports from standalone entry
- Move standalone server bootstrap to main.ts (new entry point)
- Keep index.ts as library re-exports only (for @gruenerator/api)
- Update Dockerfile CMD, package.json scripts to use main.ts/main.js
